app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
asgiref==3.8.1
    # via django
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
django==5.0.6
    # via
    #   -r .\requirements.in
    #   django-jazzmin
    #   djangorestframework
django-jazzmin==3.0.0
    # via -r .\requirements.in
djangorestframework==3.15.1
    # via -r .\requirements.in
idna==3.7
    # via requests
requests==2.32.3
    # via -r .\requirements.in
sqlparse==0.5.0
    # via django
tzdata==2024.1
    # via django
urllib3==2.2.2
    # via requests
